requirements.txt
logfire
python-dotenv==1.2.1
requests==2.32.5
openai==2.8.0
instructor==1.13.0
replicate==1.0.7
elevenlabs==2.22.1
transformers==4.50.1
sentence-transformers==3.1.1
pillow==12.0.0
opencv-python==4.10.0.84
pytesseract==0.3.10
pydantic==2.12.4
torch==2.8.0
pexels-api==1.0.1
